SawTrap
Changed the saw trap collision to the trigger and optimized it.

+++ Assets/Scripts/Character/HealthBar.cs
@@ -6,12 +6,12 @@ using UnityEngine.UI;
 public class HealthBar : MonoBehaviour
 {
     public Slider slider;
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
-    public void setHealth(int health)
+    public void setHealth(float health)
     {
     slider.value = health; 
     }
